ci(github): split git mirror workflows
Splitting Git Mirrow Action Workflows to enable individual syncing to different VCS.
This is because the underlying Github Action sets the git remote as 'target' which then the next step
tries to set again & fails. Therefore syncing is done to Gitlab, but never done to Bitbucket as the
'target' remote can not be set twice with the same name but different value.

Splitting this into separate workflows should allow for proper syncing into different VCS as the actions run on different containers allowing for setting the same remote name on different envs.