feat(algorithms): arrays - majority element with sorting
Finds the majority element using sorting using the intuition that the element
that occurs more than n/2 times in a list(where n is the size of the list) will
occupy the n/2 position of a sorted list.

The assumption made here is that the list can be sorted in place. This incurs
a time complexity of O(n log(n)) since sorting is invocled and space complexity
of O(1) since no extra space is required.