feat(puzzles): find bitonic peak using binary search
This adds a new function to find a bitonic peak in a given list of integers.
This is a variation of find_peak_element and achieves the same outcome with the same
time and space complexity of O(1) and O(log(n)).

Additional tests and comments have been added for clarity to make the algorithm easier to
understand and reason about.