feat(puzzles): find first entry in list of duplicates
Adds a new function to find the first entry in a list of duplicates using both the linear
approach and using binary search approach.

The linear approach is straightforward and outlined in the code clearly. The binary search
approach uses the same approach of a binary search with the only difference being in how
the first occurrence of the target element is obtained from the sorted list(sorted in ascending order).