feat(algorithms, array): is valid subsequence
Adds the validate subsequence algorithm solution to check if an array
is a valid subsequence of another input array. Here integers are used.
However, this can be modified to take in generics to where the types are
comparable to check if one is a valid subsequence of another input list.

This provides two approaches to the problem. One iterates through the
main array checking if the first element in the subsequence equals to
the first element in the array and if true, moves the `seek` pointer up
by 1.

If not, the `seek` pointer is checked against the length of the
subsequence to check if it has reached the end and if so, then the
sequence is a valid subsequence of the input array.

Similar approach is applied to the second alternative solution. However,
two pointers are used, were `i` is the pointer that will be used on the
subsequence and `j` is the pointer that will be used on the main input
array. The `while` loop condition is used to ensure that both `i` and
`j` do not go over the bounds and checks if the value at pointer `i` on
the sequence is equal to the value at pointer `j` on the input array. If
so, then the `i` pointer is moved up by one. Either way, the `j` pointer
is moved up by one along the main input array.
After the loop complets, the `i` pointer is checked against the length
of the sequence to verify that it reached the end. if not, then the
sequence is not a valid subsequence otherwise it is regarded as a valid
subsequence.