feat(data-structures, trie, stream-checker): adds a stream checker data structure for suffice matches
The stream checker data structure leverages the use of the Trie data
structure to find suffixes that match words that it was initialized
with. Since the trie is a Prefix Tree essentially matching on prefixes,
this required a reverse of the Trie to instead match on suffixes.

Note that not change to the Trie node is changed other than the
initialization of using char in the constructor.

BREAKING CHANGE

The Trie data structure does not handle the search
correctly anymore and will need to be refactored to cater for the
changes that have been introduced.