feat(strings, trie): is prefix of word
This adds a function `is_prefix_of_word` that enables searching for the
index of a `search_word` in a given sentence and return the index if it
exists. The index is 1-based.

This modifies the `TrieNode` allowing callers to use the `TrieNode` to
build a trie from the node and perform prefix searches a well as
modifies the `Trie` class as well adding modifications and optimizations
that match the changes in the `TrieNode` class.

Also note that this change includes an `index` to the `TrieNode` to
track the earliest index of a given character in a sentence. That is,
what is the earliest that a word has been seen, This allows returning
multiple words in the case of using the `Trie` `search` functionality.